proyectoChunche/requirements.txt
django == 3.2.18
djangorestframework == 3.14.0
django-environ==0.10.0
django-cors-headers==3.14.0